Move DBSCAN callbacks into a separate file
